Clean Code, ftp list_genome/update_refseq/update_genbank

+++ gtdb_migration_tk/ftp_manager_tools.py
@@ -27,10 +27,11 @@ import tempfile
 from datetime import datetime
 import multiprocessing as mp
 import logging
+from tqdm import tqdm
 
 
 class FTPTools(object):
-    def __init__(self, report, genomes_to_review, genome_domain_dict):
+    def __init__(self, report, genomes_to_review, genome_domain_dict,dry_run):
         self.reports = ("_assembly_report.txt",
                         "_assembly_stats.txt", "_hashes.txt")
 
@@ -49,7 +50,7 @@ class FTPTools(object):
         self.report = report
         self.genomes_to_review = genomes_to_review
         self.genome_domain_dict = genome_domain_dict
-        self.dry_run = True
+        self.dry_run = dry_run
 
     def rreplace(self, s, old, new, occurrence):
         '''
@@ -92,9 +93,7 @@ class FTPTools(object):
         try:
             workerProc = [mp.Process(target=self.__workerThread, args=(
                 workerQueue, writerQueue)) for _ in range(threads)]
-            writeProc = mp.Process(target=self.__writerThread, args=(
-                len(intersect_list), writerQueue))
-
+            writeProc = mp.Process(target=self.__listener, args=(len(intersect_list), writerQueue))
             writeProc.start()
 
             for p in workerProc:
@@ -123,22 +122,25 @@ class FTPTools(object):
                 gtdb_dir, ftp_dir, target_dir, gca_record)
             queueOut.put(status_gca)
 
-    def __writerThread(self, numgenometoprocess, writerQueue):
-        """Store or write results of worker threads in a single thread."""
-        processedItems = 0
-        while True:
-            a = writerQueue.get(block=True, timeout=None)
-            if a is None:
-                break
-            self.report.write(a)
+    def __listener(self,numgenometoprocess, writerQueue):
+        pbar = tqdm(total=numgenometoprocess)
+        for item in iter(writerQueue.get, None):
+            self.report.write(item)
+            pbar.update()
 
-            processedItems += 1
-            statusStr = 'Finished processing %d of %d (%.2f%%) genome pairs.' % (
-                processedItems, numgenometoprocess, float(processedItems) * 100 / numgenometoprocess)
-            sys.stdout.write('%s\r' % statusStr)
-            sys.stdout.flush()
+    # def __writerThread(self, numgenometoprocess, writerQueue):
+    #     """Store or write results of worker threads in a single thread."""
+    #     processedItems = 0
+    #     while True:
+    #         a = writerQueue.get(block=True, timeout=None)
+    #         if a is None:
+    #             break
+    #         self.report.write(a)
+    #
+    #         processedItems += 1
+    #         print('Finished processing %d of %d (%.2f%%) genome pairs.' % (
+    #             processedItems, numgenometoprocess, float(processedItems) * 100 / numgenometoprocess),end='\r')
 
-        sys.stdout.write('\n')
 
     def addGenomes(self, added_dict, ftp_dir, new_directory, genome_domain_dict):
         '''
@@ -155,32 +157,25 @@ class FTPTools(object):
         :param new_directory:base directory leading the new repository for refseq
         '''
 
-        processedItems = 0
-        for gcf_record in added_dict:
-            processedItems += 1
-            statusStr = 'Finished processing %d of %d (%.2f%%) genome to add.' % (
-                processedItems, len(added_dict), float(processedItems) * 100 / len(added_dict))
-            sys.stdout.write('%s\r' % statusStr)
-            sys.stdout.flush()
+        for gcf_record,path_record in tqdm(added_dict.items()):
 
             target_dir = os.path.join(
-                new_directory, added_dict[gcf_record].replace(ftp_dir, ''))
-            shutil.copytree(added_dict[
-                gcf_record], target_dir, ignore=shutil.ignore_patterns("*_assembly_structure"))
+                new_directory, path_record.replace(ftp_dir, ''))
             self.report.write(
                 "{0}\t{1}\tnew\n".format(genome_domain_dict.get(gcf_record).upper(), gcf_record))
-            for compressed_file in glob.glob(target_dir + "/*.gz"):
-                if not os.path.isdir(compressed_file):
-                    inF = gzip.open(compressed_file, 'rb')
-                    outF = open(
-                        self.rreplace(compressed_file, ".gz", "", 1), 'wb')
-                    outF.write(inF.read())
-                    inF.close()
-                    outF.close()
-                    os.remove(compressed_file)
-        sys.stdout.write('\n')
+            if not self.dry_run:
+                shutil.copytree(path_record, target_dir, ignore=shutil.ignore_patterns("*_assembly_structure"))
+                for compressed_file in glob.glob(target_dir + "/*.gz"):
+                    if not os.path.isdir(compressed_file):
+                        inF = gzip.open(compressed_file, 'rb')
+                        outF = open(
+                            self.rreplace(compressed_file, ".gz", "", 1), 'wb')
+                        outF.write(inF.read())
+                        inF.close()
+                        outF.close()
+                        os.remove(compressed_file)
 
-    def removeGenomes(self, removed_dict, old_dict_domain):
+    def removeGenomes(self, removed_dict):
         '''
         removeGenomes function removes all outdated genomes from the gtdb database
         In addition it tracks the lists(name and owner) that have been modified while deleting those genomes
@@ -190,7 +185,7 @@ class FTPTools(object):
 
         for gca_record in removed_dict:
             self.report.write(
-                "{0}\t{1}\tremoved\n".format(old_dict_domain.get(gca_record).upper(), gca_record))
+                "UNDEFINED\t{0}\tremoved\n".format(gca_record))
 
     def readmd5Checksum(self, gtdb_dir, ftp_dir, target_dir, genome_record):
         '''
@@ -200,132 +195,133 @@ class FTPTools(object):
         pathgtdbmd5 = os.path.join(gtdb_dir, "md5checksums.txt")
         target_pathnewmd5 = os.path.join(target_dir, "md5checksums.txt")
         status = []
-
-        tmp_ftp_dir = tempfile.mkdtemp()
-        tmp_target = os.path.join(tmp_ftp_dir, os.path.basename(target_dir))
-        shutil.copytree(ftp_dir, tmp_target, symlinks=True,
-                        ignore=shutil.ignore_patterns("*_assembly_structure"))
-        for compressed_file in glob.glob(tmp_target + "/*.gz"):
-            if os.path.isdir(compressed_file) is False:
-                inF = gzip.open(compressed_file, 'rb')
-                try:
-                    outF = open(
-                        self.rreplace(compressed_file, ".gz", "", 1), 'wb')
-                except IOError:
-                    os.chmod(
-                        self.rreplace(compressed_file, ".gz", "", 1), 0o775)
-                    outF = open(
-                        self.rreplace(compressed_file, ".gz", "", 1), 'wb')
-                outF.write(inF.read())
-                inF.close()
-                outF.close()
-                os.remove(compressed_file)
-
-        ftpdict, ftpdict_fasta, ftpdict_faa, ftpdict_extra_fasta = self.parse_checksum(
-            tmp_target)
-        gtdbdict, gtdbdict_fasta, gtdbdict_faa, gtdbdict_extra_fasta = self.parse_checksum(
-            gtdb_dir)
-
-        # if the genomic.fna.gz or the protein.faa.gz are missing, we set this
-        # record as incomplete
-        if len(list(set(ftpdict_fasta.keys()).symmetric_difference(set(gtdbdict_fasta.keys())))) > 0:
-            self.genomes_to_review.write("tmp_target:{}\nftpdict_fasta.keys():{}\ngtdb_dir:{}\ngtdbdict_fasta.keys():{}\n\n".format(tmp_target,
-                                                                                                                                    ftpdict_fasta.keys(),
-                                                                                                                                    gtdb_dir,
-                                                                                                                                    gtdbdict_fasta.keys()))
-            status.append("incomplete")
-            shutil.copytree(tmp_target, target_dir, symlinks=True,
+        if not self.dry_run:
+            tmp_ftp_dir = tempfile.mkdtemp()
+            tmp_target = os.path.join(tmp_ftp_dir, os.path.basename(target_dir))
+            shutil.copytree(ftp_dir, tmp_target, symlinks=True,
                             ignore=shutil.ignore_patterns("*_assembly_structure"))
-            # we unzip of gz file
+            for compressed_file in glob.glob(tmp_target + "/*.gz"):
+                if os.path.isdir(compressed_file) is False:
+                    inF = gzip.open(compressed_file, 'rb')
+                    try:
+                        outF = open(
+                            self.rreplace(compressed_file, ".gz", "", 1), 'wb')
+                    except IOError:
+                        os.chmod(
+                            self.rreplace(compressed_file, ".gz", "", 1), 0o775)
+                        outF = open(
+                            self.rreplace(compressed_file, ".gz", "", 1), 'wb')
+                    outF.write(inF.read())
+                    inF.close()
+                    outF.close()
+                    os.remove(compressed_file)
 
-        else:
-            ftp_folder = False
-            # check if genomic.fna.gz and protein.faa.gz are similar between
-            # previous gtdb and ftp
-            for key, value in ftpdict_fasta.items():
-                if value != gtdbdict_fasta.get(key):
-                    ftp_folder = True
+            ftpdict, ftpdict_fasta, ftpdict_faa, ftpdict_extra_fasta = self.parse_checksum(
+                tmp_target)
+            gtdbdict, gtdbdict_fasta, gtdbdict_faa, gtdbdict_extra_fasta = self.parse_checksum(
+                gtdb_dir)
 
-            # if one of the 2 files is different than the previous version , we
-            # use the ftp record over the previous gtdb one , we then need to
-            # re run the metadata generation
-            if ftp_folder:
-                if os.path.exists(target_dir):
-                    shutil.rmtree(target_dir)
-                shutil.copytree(
-                    tmp_target, target_dir, symlinks=True,
-                    ignore=shutil.ignore_patterns("*_assembly_structure"))
-                status.append("modified")
+            # if the genomic.fna.gz or the protein.faa.gz are missing, we set this
+            # record as incomplete
+            if len(list(set(ftpdict_fasta.keys()).symmetric_difference(set(gtdbdict_fasta.keys())))) > 0:
+                self.genomes_to_review.write("tmp_target:{}\nftpdict_fasta.keys():{}\ngtdb_dir:{}\ngtdbdict_fasta.keys():{}\n\n".format(tmp_target,
+                                                                                                                                        ftpdict_fasta.keys(),
+                                                                                                                                        gtdb_dir,
+                                                                                                                                        gtdbdict_fasta.keys()))
+                status.append("incomplete")
+                shutil.copytree(tmp_target, target_dir, symlinks=True,
+                                ignore=shutil.ignore_patterns("*_assembly_structure"))
+                # we unzip of gz file
 
             else:
-                # The 2 main fasta files haven't changed so we can copy the old
-                # gtdb folder over
-                if os.path.exists(target_dir):
-                    shutil.rmtree(target_dir)
-                shutil.copytree(
-                    gtdb_dir, target_dir, symlinks=True,
-                    ignore=shutil.ignore_patterns("*_assembly_structure"))
-                if not os.path.isdir(os.path.join(target_dir,'prodigal')):
-                    status.append("unmodified")
+                ftp_folder = False
+                # check if genomic.fna.gz and protein.faa.gz are similar between
+                # previous gtdb and ftp
+                for key, value in ftpdict_fasta.items():
+                    if value != gtdbdict_fasta.get(key):
+                        ftp_folder = True
+
+                # if one of the 2 files is different than the previous version , we
+                # use the ftp record over the previous gtdb one , we then need to
+                # re run the metadata generation
+                if ftp_folder:
+                    if os.path.exists(target_dir):
+                        shutil.rmtree(target_dir)
+                    shutil.copytree(
+                        tmp_target, target_dir, symlinks=True,
+                        ignore=shutil.ignore_patterns("*_assembly_structure"))
+                    status.append("modified")
+
+                else:
+                    # The 2 main fasta files haven't changed so we can copy the old
+                    # gtdb folder over
+                    if os.path.exists(target_dir):
+                        shutil.rmtree(target_dir)
+                    shutil.copytree(
+                        gtdb_dir, target_dir, symlinks=True,
+                        ignore=shutil.ignore_patterns("*_assembly_structure"))
+                    if not os.path.isdir(os.path.join(target_dir,'prodigal')):
+                        status.append("unmodified")
 
 
 
-                # We check if all other file of this folder are the same.
-                checksum_changed = False
+                    # We check if all other file of this folder are the same.
+                    checksum_changed = False
 
-                for key, value in ftpdict_faa.items():
-                    if value != gtdbdict_faa.get(key):
-                        checksum_changed = True
-                        shutil.copy2(
-                            os.path.join(tmp_target, key), os.path.join(target_dir, key))
-                        status.append("new_protein")
+                    for key, value in ftpdict_faa.items():
+                        if value != gtdbdict_faa.get(key):
+                            checksum_changed = True
+                            shutil.copy2(
+                                os.path.join(tmp_target, key), os.path.join(target_dir, key))
+                            status.append("new_protein")
 
-                for key, value in ftpdict.items():
-                    if value != gtdbdict.get(key):
-                        checksum_changed = True
-                        shutil.copy2(
-                            os.path.join(tmp_target, key), os.path.join(target_dir, key))
-                        status.append("new_metadata")
+                    for key, value in ftpdict.items():
+                        if value != gtdbdict.get(key):
+                            checksum_changed = True
+                            shutil.copy2(
+                                os.path.join(tmp_target, key), os.path.join(target_dir, key))
+                            status.append("new_metadata")
 
-                for key, value in ftpdict_extra_fasta.items():
-                    if value != gtdbdict_extra_fasta.get(key):
-                        checksum_changed = True
-                        shutil.copy2(
-                            os.path.join(tmp_target, key), os.path.join(target_dir, key))
-                        status.append("new_cds_rna")
-                # we copy the new checksum
-                if checksum_changed:
-                    try:
-                        shutil.copy2(pathgtdbmd5, target_pathnewmd5)
-                    except IOError:
-                        os.chmod(target_pathnewmd5, 0o664)
-                        shutil.copy2(pathftpmd5, target_pathnewmd5)
-                for report in self.reports:
-                    target_files = glob.glob(
-                        os.path.join(target_dir, "*" + report))
-                    ftp_files = glob.glob(os.path.join(ftp_dir, "*" + report))
-                    if len(target_files) == 1 and len(ftp_files) == 1:
-                        status = self.comparesha256(
-                            ftp_files[0], target_files[0], status)
-                    elif len(target_files) == 0 and len(ftp_files) == 0 and report == '_hashes.txt':
-                        status.append("old_folder_dir")
-                    elif len(target_files) == 0 and len(ftp_files) == 1 and report == '_hashes.txt':
-                        shutil.copy2(ftp_dir[0], ftp_dir[0].replace(
-                            ftp_dir.target_dir))
-                        status.append("new_hashes")
-                    else:
-                        print("########")
-                        print(target_dir)
-                        print(target_files)
-                        print(ftp_dir)
-                        print(ftp_files)
-                        print(f"IT SHOULDN'T HAPPEN ({target_dir},{ftp_dir}) ")
-                        print("########")
-                        status.append("to_curate")
-        status_record = "{0}\t{1}\t{2}\n".format(self.genome_domain_dict.get(
-            genome_record).upper(), genome_record, ';'.join([x for x in set(status)]))
-        shutil.rmtree(tmp_ftp_dir)
-        return status_record
+                    for key, value in ftpdict_extra_fasta.items():
+                        if value != gtdbdict_extra_fasta.get(key):
+                            checksum_changed = True
+                            shutil.copy2(
+                                os.path.join(tmp_target, key), os.path.join(target_dir, key))
+                            status.append("new_cds_rna")
+                    # we copy the new checksum
+                    if checksum_changed:
+                        try:
+                            shutil.copy2(pathgtdbmd5, target_pathnewmd5)
+                        except IOError:
+                            os.chmod(target_pathnewmd5, 0o664)
+                            shutil.copy2(pathftpmd5, target_pathnewmd5)
+                    for report in self.reports:
+                        target_files = glob.glob(
+                            os.path.join(target_dir, "*" + report))
+                        ftp_files = glob.glob(os.path.join(ftp_dir, "*" + report))
+                        if len(target_files) == 1 and len(ftp_files) == 1:
+                            status = self.comparesha256(
+                                ftp_files[0], target_files[0], status)
+                        elif len(target_files) == 0 and len(ftp_files) == 0 and report == '_hashes.txt':
+                            status.append("old_folder_dir")
+                        elif len(target_files) == 0 and len(ftp_files) == 1 and report == '_hashes.txt':
+                            shutil.copy2(ftp_dir[0], ftp_dir[0].replace(
+                                ftp_dir.target_dir))
+                            status.append("new_hashes")
+                        else:
+                            print("########")
+                            print(target_dir)
+                            print(target_files)
+                            print(ftp_dir)
+                            print(ftp_files)
+                            print(f"IT SHOULDN'T HAPPEN ({target_dir},{ftp_dir}) ")
+                            print("########")
+                            status.append("to_curate")
+            status_record = "{0}\t{1}\t{2}\n".format(self.genome_domain_dict.get(
+                genome_record).upper(), genome_record, ';'.join([x for x in set(status)]))
+            shutil.rmtree(tmp_ftp_dir)
+            return status_record
+        return "{0}\t{1}\t{2}\n".format(self.genome_domain_dict.get(genome_record).upper(), genome_record,'undefined comparison')
 
     def comparesha256(self, ftp_file, target_file, status):
         '''
